refactor(types): move custom `@types` packages
These were in module-scan but really belong as their own packages.
