+++ apps/module-scan/package.json
@@ -58,18 +58,19 @@
     "@types/body-parser": "^1.19.0",
     "@types/busboy": "^0.2.3",
     "@types/debug": "^4.1.5",
+    "@types/deep-eql": "workspace:*",
     "@types/express": "^4.16.1",
     "@types/fs-extra": "^9.0.6",
     "@types/jest": "^26.0.6",
     "@types/multer": "^1.4.3",
     "@types/node": "^12.0.0",
     "@types/pdfjs-dist": "^2.1.3",
-    "@types/readable-stream": "^2.3.9",
     "@types/sqlite3": "^3.1.6",
     "@types/supertest": "^2.0.10",
     "@types/tmp": "^0.2.0",
     "@types/uuid": "^8.3.0",
     "@types/yauzl": "^2.9.1",
+    "@types/zip-stream": "workspace:*",
     "@typescript-eslint/eslint-plugin": "^4.13.0",
     "@typescript-eslint/parser": "^4.13.0",
     "chalk": "^4.1.0",
