--- apps/module-scan/types/deep-eql.d.ts
@@ -1,3 +1,0 @@
-declare module 'deep-eql' {
-  export default function deepEql(a: unknown, b: unknown): boolean
-}
