+++ libs/@types/deep-eql/deep-eql.d.ts
@@ -1,0 +1,1 @@
+export default function deepEql(a: unknown, b: unknown): boolean
