+++ libs/@types/deep-eql/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@types/deep-eql",
+  "version": "0.0.1-development",
+  "types": "deep-eql"
+}
